Finish renaming `draft-state-for` to `draft-for`
