+++ tests/dummy/app/templates/components/profile.hbs
@@ -1,2 +1,2 @@
 <CurrentProfile @user={{this.userInfo}} />
-<ProfileForm @user={{draft-state-for this.userInfo}} />
+<ProfileForm @user={{draft-for this.userInfo}} />
